Refactored spaceship to support soon support iTC
