+++ lib/spaceship/developer_portal/developer_portal_client.rb
@@ -1,0 +1,257 @@
+module Spaceship
+  class DeveloperPortalClient < Spaceship::Client
+
+    #####################################################
+    # @!group Init and Login
+    #####################################################
+
+    def self.hostname
+      "https://developer.apple.com/services-account/#{PROTOCOL_VERSION}/"
+    end
+
+    # Fetches the latest API Key from the Apple Dev Portal
+    def api_key
+      cache_path = "/tmp/spaceship_api_key.txt"
+      begin
+        cached = File.read(cache_path) 
+      rescue Errno::ENOENT
+      end
+      return cached if cached
+
+      landing_url = "https://developer.apple.com/membercenter/index.action"
+      logger.info("GET: " + landing_url)
+      headers = @client.get(landing_url).headers
+      results = headers['location'].match(/.*appIdKey=(\h+)/)
+      if results.length > 1
+        api_key = results[1]
+        File.write(cache_path, api_key)
+        return api_key
+      else
+        raise "Could not find latest API Key from the Dev Portal"
+      end
+    end
+
+    def send_login_request(user, password)
+      response = request(:post, "https://idmsa.apple.com/IDMSWebAuth/authenticate", {
+        appleId: user,
+        accountPassword: password,
+        appIdKey: api_key
+      })
+
+      if response['Set-Cookie'] =~ /myacinfo=(\w+);/
+        @cookie = "myacinfo=#{$1};"
+        return @client
+      else
+        # User Credentials are wrong
+        raise InvalidUserCredentialsError.new(response)
+      end
+    end
+
+    def teams
+      r = request(:post, 'account/listTeams.action')
+      parse_response(r, 'teams')
+    end
+
+    def select_team
+      @current_team_id = self.UI.select_team
+    end
+
+
+    #####################################################
+    # @!group Apps
+    #####################################################
+
+    def apps
+      paging do |page_number|
+        r = request(:post, 'account/ios/identifiers/listAppIds.action', {
+          teamId: team_id,
+          pageNumber: page_number,
+          pageSize: page_size,
+          sort: 'name=asc'
+        })
+        parse_response(r, 'appIds')
+      end
+    end
+
+    def details_for_app(app)
+      r = request(:post, 'account/ios/identifiers/getAppIdDetail.action', {
+        teamId: team_id,
+        appIdId: app.app_id
+      })
+      parse_response(r, 'appId')
+    end
+
+    def create_app!(type, name, bundle_id)
+      ident_params = case type.to_sym
+      when :explicit
+        {
+          type: 'explicit',
+          explicitIdentifier: bundle_id,
+          appIdentifierString: bundle_id,
+          push: 'on',
+          inAppPurchase: 'on',
+          gameCenter: 'on'
+        }
+      when :wildcard
+        {
+          type: 'wildcard',
+          wildcardIdentifier: bundle_id,
+          appIdentifierString: bundle_id
+        }
+      end
+
+      params = {
+        appIdName: name,
+        teamId: team_id
+      }
+
+      params.merge!(ident_params)
+
+      r = request(:post, 'account/ios/identifiers/addAppId.action', params)
+      parse_response(r, 'appId')
+    end
+
+    def delete_app!(app_id)
+      r = request(:post, 'account/ios/identifiers/deleteAppId.action', {
+        teamId: team_id,
+        appIdId: app_id
+      })
+      parse_response(r)
+    end
+
+    #####################################################
+    # @!group Devices
+    #####################################################
+
+    def devices
+      paging do |page_number|
+        r = request(:post, 'account/ios/device/listDevices.action', {
+          teamId: team_id,
+          pageNumber: page_number,
+          pageSize: page_size,
+          sort: 'name=asc'
+        })
+        parse_response(r, 'devices')
+      end
+    end
+
+    def create_device!(device_name, device_id)
+      r = request(:post) do |r|
+        r.url "https://developerservices2.apple.com/services/#{PROTOCOL_VERSION}/ios/addDevice.action"
+        r.params = {
+          teamId: team_id,
+          deviceNumber: device_id,
+          name: device_name
+        }
+      end
+
+      parse_response(r, 'device')
+    end
+
+    #####################################################
+    # @!group Certificates
+    #####################################################
+
+    def certificates(types)
+      paging do |page_number|
+        r = request(:post, 'account/ios/certificate/listCertRequests.action', {
+          teamId: team_id,
+          types: types.join(','),
+          pageNumber: page_number,
+          pageSize: page_size,
+          sort: 'certRequestStatusCode=asc'
+        })
+        parse_response(r, 'certRequests')
+      end
+    end
+
+    def create_certificate!(type, csr, app_id = nil)
+      r = request(:post, 'account/ios/certificate/submitCertificateRequest.action', {
+        teamId: team_id,
+        type: type,
+        csrContent: csr,
+        appIdId: app_id  #optional
+      })
+      parse_response(r, 'certRequest')
+    end
+
+    def download_certificate(certificate_id, type)
+      {type: type, certificate_id: certificate_id}.each { |k, v| raise "#{k} must not be nil" if v.nil? }
+
+      r = request(:post, 'https://developer.apple.com/account/ios/certificate/certificateContentDownload.action', {
+        teamId: team_id,
+        displayId: certificate_id,
+        type: type
+      })
+      parse_response(r)
+    end
+
+    def revoke_certificate!(certificate_id, type)
+      r = request(:post, 'account/ios/certificate/revokeCertificate.action', {
+        teamId: team_id,
+        certificateId: certificate_id,
+        type: type
+      })
+      parse_response(r, 'certRequests')
+    end
+
+    #####################################################
+    # @!group Provisioning Profiles
+    #####################################################
+
+    def provisioning_profiles
+      r = request(:post) do |r|
+        r.url "https://developerservices2.apple.com/services/#{PROTOCOL_VERSION}/ios/listProvisioningProfiles.action"
+        r.params = {
+          teamId: team_id,
+          includeInactiveProfiles: true,
+          onlyCountLists: true,
+        }
+      end
+
+      parse_response(r, 'provisioningProfiles')
+    end
+
+    def create_provisioning_profile!(name, distribution_method, app_id, certificate_ids, device_ids)
+      r = request(:post, 'account/ios/profile/createProvisioningProfile.action', {
+        teamId: team_id,
+        provisioningProfileName: name,
+        appIdId: app_id,
+        distributionType: distribution_method,
+        certificateIds: certificate_ids,
+        deviceIds: device_ids
+      })
+      parse_response(r, 'provisioningProfile')
+    end
+
+    def download_provisioning_profile(profile_id)
+      r = request(:get, 'https://developer.apple.com/account/ios/profile/profileContentDownload.action', {
+        teamId: team_id,
+        displayId: profile_id
+      })
+      parse_response(r)
+    end
+
+    def delete_provisioning_profile!(profile_id)
+      r = request(:post, 'account/ios/profile/deleteProvisioningProfile.action', {
+        teamId: team_id,
+        provisioningProfileId: profile_id
+      })
+      parse_response(r)
+    end
+
+    def repair_provisioning_profile!(profile_id, name, distribution_method, app_id, certificate_ids, device_ids)
+      r = request(:post, 'account/ios/profile/regenProvisioningProfile.action', {
+        teamId: team_id,
+        provisioningProfileId: profile_id,
+        provisioningProfileName: name,
+        appIdId: app_id,
+        distributionType: distribution_method,
+        certificateIds: certificate_ids.first, # we are most of the times only allowed to pass one
+        deviceIds: device_ids
+      })
+
+      parse_response(r, 'provisioningProfile')
+    end
+  end
+end
