+++ lib/spaceship/developer_portal/portal_base.rb
@@ -1,0 +1,9 @@
+module Spaceship
+  class PortalBase < Spaceship::Base
+    class << self
+      def client
+        @client || Spaceship::Portal.client
+      end
+    end
+  end
+end
