+++ lib/spaceship/developer_portal/provisioning_profile.rb
@@ -1,6 +1,6 @@
 module Spaceship
   # Represents a provisioning profile of the Apple Dev Portal
-  class ProvisioningProfile < Base
+  class ProvisioningProfile < PortalBase
     # @return (String) The ID generated by the Dev Portal
     #   You'll probably not really need this value
     # @example
