+++ lib/spaceship/developer_portal/ui/select_team.rb
